refactoring: move spotify component on it's own folder
