+++ pages/home/track.tsx
@@ -11,7 +11,7 @@ import { Layout, Loader, NoData } from "@components/index";
 import styles from "@styles/Pages/artist.module.scss";
 
 const Error = dynamic(async () => await import("next/error"));
-const TrackCard = dynamic(async () => await import("@components/TrackCard"), {
+const TrackCard = dynamic(async () => await import("@components/Spotify/TrackCard"), {
   suspense: true,
 });
 const Pagination = dynamic(async () => await import("@components/Pagination").then((res) => res.Pagination));
